+++ docs/source/changelog.rst
@@ -1577,6 +1577,8 @@ Changes
   - The NLP cloud client checks that all requested processors are available
     remotely, and fails overtly, rather than silently ignoring those.
 
+  - Alcohol units regex NLP.
+
 - Bayesian linkage, identifiable and de-identified.
 
 - **Minimum Python version is now 3.8.** (Required by numpy v1.22.)